# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GAC授权中心",
        "title": "GAC授权中心",
        "termsOfService": "https://paymedia.yuque.com/os/hb9uhc",
        "contact": {
            "name": "org.pm.gac.service",
            "url": "https://www.apifox.cn/web/main"
        },
        "license": {
            "name": "GIN.swagger",
            "url": "https://github.com/swaggo/swag/blob/master/README_zh-CN.md#%E5%A6%82%E4%BD%95%E4%B8%8Egin%E9%9B%86%E6%88%90"
        },
        "version": "2.0"
    },
    "host": "api.t.dev.pay.fun",
    "basePath": "/gac/v1",
    "paths": {
...
        "/sso/auth/my_permission": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO/AUTH"
                ],
                "summary": "我的权限（菜单+按钮）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.PermissionAuthReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ssoApi.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/resp.SsoMenuResps"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    }
                }
            }
        },
        "/sso/auth/passport": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO/AUTH"
                ],
                "summary": "Passport认证（后端服务之间调用）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.SsoPassportAuthReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    }
                }
            }
        },
        "/sso/auth/platform/info": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO/AUTH"
                ],
                "summary": "我的信息（平台）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ssoApi.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/assembler.AccountAdmin"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    }
                }
            }
        },
        "/sso/auth/platform/update": {
            "post": {
                "description": "更新",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO/AUTH"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateAccountAdmin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data=OK",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    }
                }
            }
        },
        "/sso/auth/platform/update_password": {
            "post": {
                "description": "更新密码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO/AUTH"
                ],
                "summary": "更新密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AcctChangePasswd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data=OK",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    }
                }
            }
        },
...
        "/sso/auth/subclass/info": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO/AUTH"
                ],
                "summary": "我的信息（子账户）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    }
                }
            }
        },
        "/sso/auth/subclass/rest_password": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO/AUTH"
                ],
                "summary": "子账户重置密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AuthSubclassResetPasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    }
                }
            }
        },
        "/sso/auth/subclass/update": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO/AUTH"
                ],
                "summary": "编辑我的子账户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AuthSubclassUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ssoApi.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "assembler.AccountAdmin": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "账户",
                    "type": "string"
                },
                "account_id": {
                    "description": "sso账号id",
                    "type": "string"
                },
                "account_type": {
                    "description": "账户类型",
                    "type": "integer"
                },
                "created_id": {
                    "description": "创建管理ID",
                    "type": "string"
                },
                "created_ip": {
                    "description": "创建Ip",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "created_user": {
                    "type": "string"
                },
                "creator_account_type": {
                    "description": "创建人账户类别(A:超管 B:管理员 C:员工)",
                    "type": "string"
                },
                "creator_name": {
                    "description": "创建人名称",
                    "type": "string"
                },
                "is_admin": {
                    "description": "是否是超级管理员：1:是 0 否",
                    "type": "integer"
                },
                "last_login_ip": {
                    "description": "最后登陆ip",
                    "type": "string"
                },
                "last_login_time": {
                    "description": "最后登陆时间",
                    "type": "string"
                },
                "login_count": {
                    "description": "登陆次数",
                    "type": "integer"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "password": {
                    "description": "登陆密码",
                    "type": "string"
                },
                "pk": {
                    "type": "integer"
                },
                "role_name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "role_pk": {
                    "description": "角色id",
                    "type": "integer"
                },
                "role_type": {
                    "description": "角色类型",
                    "type": "string"
                },
                "secret": {
                    "description": "安全秘钥",
                    "type": "string"
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "req.AcctChangePasswd": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
...
        "req.AuthSubclassResetPasswordReq": {
            "type": "object",
            "required": [
                "old_password",
                "password"
            ],
            "properties": {
                "old_password": {
                    "description": "旧密码",
                    "type": "string"
                },
                "password": {
                    "description": "新密码",
                    "type": "string"
                }
            }
        },
        "req.AuthSubclassUpdateReq": {
            "type": "object",
            "properties": {
                "mobile": {
                    "description": "电话",
                    "type": "string"
                },
                "name": {
                    "description": "账户名称",
                    "type": "string"
                }
            }
        },
...
        "req.PermissionAuthReq": {
            "type": "object",
            "properties": {
                "apps": {
                    "description": "应用CODE, GAC子系统默认单选:GAC",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "orgs": {
                    "description": "组织CODE",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "req.SsoPassportAuthReq": {
            "type": "object",
            "required": [
                "method",
                "uri"
            ],
            "properties": {
                "method": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "req.UpdateAccountAdmin": {
            "type": "object",
            "properties": {
                "account_id": {
                    "description": "账户ID 更新时使用",
                    "type": "string"
                },
                "account_name": {
                    "description": "账号名称",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string"
                },
                "role_pk": {
                    "description": "角色主键值",
                    "type": "integer"
                }
            }
        },
...
        "resp.AuthAdminMenuResp": {
            "type": "object",
            "properties": {
                "api_route": {
                    "description": "后端接口路由",
                    "type": "string"
                },
                "code": {
                    "description": "code",
                    "type": "string"
                },
                "i18n_code": {
                    "description": "多语言code",
                    "type": "string"
                },
                "icon": {
                    "description": "icon",
                    "type": "string"
                },
                "keep_alive": {
                    "description": "缓存 1:启用 2：禁用",
                    "type": "integer"
                },
                "method": {
                    "description": "请求方法:GET,POST,PUT,PATCH,DELETE",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parent_name": {
                    "description": "上级名称",
                    "type": "string"
                },
                "parent_pk": {
                    "type": "integer"
                },
                "pk": {
                    "type": "integer"
                },
                "redirect": {
                    "description": "前端路由重定向",
                    "type": "string"
                },
                "route": {
                    "description": "前端路由",
                    "type": "string"
                },
                "show": {
                    "description": "1:显示 2：隐藏",
                    "type": "integer"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "type": {
                    "description": "类型： 1：菜单 3：按钮",
                    "type": "integer"
                },
                "view_path": {
                    "description": "前端文件路径",
                    "type": "string"
                }
            }
        },
...
        "resp.SsoMenuResps": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/resp.SsoResp"
            }
        },
        "resp.SsoResp": {
            "type": "object",
            "properties": {
                "menu": {
                    "description": "Path string              `json:\"path\"`",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resp.AuthAdminMenuResp"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "ssoApi.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}